feat(cnchess): implement chess piece click interaction and position flags
- Add `OnChessPieceClick` method to handle piece selection and animation
- Introduce position flag template and rendering logic on the board
- Adjust layer order for chess pieces to ensure proper visual stacking
- Include utility function for sprite offset with floating point precision
- Remove obsolete assertion in `ChessAnchor2Pos`
- Add `m_bRedSide` member to support side-based logic
- Update theme configuration to include position flag assets
- Modify `ChessPiece` constructor to disable message transparency
- Add helper method `IsRedSide` to determine piece affiliation

Also updates `Util::OffsetSprite` signature in upgrade game to accept float values
for more precise animations.

+++ games/upgrade/client/SSpriteCard.cpp
@@ -242,7 +242,7 @@ void SSpriteCard::OnMouseMove(UINT nFlags, CPoint pt)
 }
 
 //------------------------------------------------------------------------
-SAutoRefPtr<IValueAnimator> Util::OffsetSprite(IWindow *pCard, int dx, int dy, int nSpeed)
+SAutoRefPtr<IValueAnimator> Util::OffsetSprite(IWindow *pCard, float dx, float dy, int nSpeed)
 {
     ILayoutParam *pParam = (ILayoutParam*)pCard->GetLayoutParam();
     SASSERT(pParam->IsClass(SAnchorLayoutParam::GetClassName()));
